backend/requirements.txt
flask
flask-cors
pymongo
dnspython
pyjwt
bcrypt
python-dotenv
requests
pypdf
python-docx
gunicorn
